% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AGD-package.R
\docType{data}
\name{boys7482}
\alias{boys7482}
\title{Growth of Dutch boys}
\format{A data frame with 7482 rows on the following 9 variables:
\describe{
\item{age}{Decimal age (0-21 years)}
\item{hgt}{Height (cm)}
\item{wgt}{Weight (kg)}
\item{bmi}{Body mass index}
\item{hc}{Head circumference (cm)}
\item{gen}{Genital Tanner stage (G1-G5)}
\item{phb}{Pubic hair (Tanner P1-P6)}
\item{tv}{Testicular volume (ml)}
\item{reg}{Region (north, east, west, south, city)}
}}
\source{
Fredriks, A.M,, van Buuren, S., Burgmeijer, R.J., Meulmeester JF,
Beuker, R.J., Brugman, E., Roede, M.J., Verloove-Vanhorick, S.P., Wit, J.M.
(2000) Continuing positive secular growth change in The Netherlands
1955-1997.  \emph{Pediatric Research}, \bold{47}, 316-323.

Fredriks, A.M., van Buuren, S., Wit, J.M., Verloove-Vanhorick, S.P. (2000).
Body index measurements in 1996-7 compared with 1980.  \emph{Archives of
Disease in Childhood}, \bold{82}, 107-112.
}
\description{
Height, weight, head circumference and puberty of 7482 Dutch boys.
}
\details{
The complete sample of cross-sectional data from boys 0-21 years used to
construct the Dutch growth references 1997. Variables \code{gen} and
\code{phb} are ordered factors. \code{reg} is a factor. Note: A 10\% sample
from this data is available in data set \code{boys} in the \code{mice}
package.
}
\author{
Stef van Buuren, 2012
}
\keyword{datasets}

