% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{References WHO}
\alias{References WHO}
\alias{who.wgt}
\title{Reference tables from WHO 2006}
\format{A data frame with seven variables: \describe{
\item{list("pop")}{Study Population} \item{list("sub")}{Subpopulation}
\item{list("sex")}{Sex (M,F)} \item{list("x")}{Decimal age (0-5 years)}
\item{list("L")}{Lambda (skewness) curve} \item{list("M")}{Median curve}
\item{list("S")}{Coefficient of Variation curve} }}
\source{
WHO Multicentre Growth Reference Study Group.  WHO Child Growth
Standards based on length/height, weight and age.  \emph{Acta Paediatr},
Suppl. 2006, 450, 76-85.
}
\description{
Reference tables from WHO 2006
}
\details{
The data were fitted by the LMS model. Parameters are stored as type
\code{LMS}. Tabulated values are point ages. The ages cover 0-5 years.

The naming conventions are as follows: \describe{
\item{list("who.wgt")}{Weight (kg) for Age} }
}
\seealso{
\code{\link{nl4.wgt}}, \code{\link{cdc.wgt}}
}
\keyword{datasets}

